nrp/planner/vqmpt/requirements_original.txt
addict==2.4.0
ansi2html==1.8.0
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
einops==0.7.0
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.9.0
Flask==3.0.0
fonttools==4.44.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
open3d==0.17.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyg-lib==0.3.0+pt20cu118
pygccxml==2.4.0
Pygments==2.16.1
pyparsing==3.1.1
pyplusplus==1.8.5
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.28.1
retrying==1.3.4
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
torch==2.0.1+cu118
torch-cluster==1.6.3+pt20cu118
torch-scatter==2.1.2+pt20cu118
torch-sparse==0.6.18+pt20cu118
torch-spline-conv==1.2.2+pt20cu118
torch_geometric==2.4.0
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.13.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.9
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.17.0